requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio-dgram==2.1.0
attrs==21.2.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
discord==1.7.3
discord.py==1.7.3
dnspython==2.1.0
idna==3.3
mcstatus==6.5.0
mctools==1.1.2
multidict==5.2.0
python-dotenv==0.19.2
six==1.16.0
typing-extensions==4.0.1
yarl==1.7.2
